perf event reordering: improve API
Before this patch, the perf events reordering code was assuming that
there was a timestamp in the first field of the perf event struct. That
convention was not explicit.

This patch removes this assumption: by default, perf events are not
ordered unless SetTimestampFunc() is called to register a callback
getting the timestamp.