Increase size of log buffer
The bpf() syscall has a log_buf parameter for the verifier to write
output. When the buffer is too small, bpf() fails, even if the program
is good otherwise.

Symptoms:
error while loading "kretprobe/foobar" (no space left on device)

Changing the size from 64K to 256K works for me in practice