elf: do object pinning with pin config
When a module is loaded with parameters of pin config, pin the object
with a path defined in the custom config. The actual pinnned object is
visible as `/sys/fs/bpf/<pin_path>`.
The pin_path can be any path, e.g. a normal path like "ns1/testgroup1",
or hash of a particular path.