bcc: only add verifier log to error string if present
If bpf_prog_load fails due to e.g. permission problems, the error string
currently looks as follows:

  $ ./my_gobpf_prog
  error loading BPF program:

which fails to inform the user that there is a permission problem. Fix
this by only showing the verifier log output if ther actually is some,
i.e. the error occurs during verification. Otherwise just show the error
string, such the the above now looks as follows:

  $ ./my_gobpf_prog
  error loading BPF program: operation not permitted

and in case of a verifier error:

  $ sudo ./my_gobpf_prog
  error loading BPF program:
  back-edge from insn 34 to 34

Signed-off-by: Tobias Klauser <tklauser@distanz.ch>