Use exact width integer as second argument type to lostCallback.  Fix #272.
`lostCallback` is exported via CGO and re-imported elsewhere as a C function.
The second argument is declared as uint64_t when re-imported, but declared
as unsigned long in the original Go function definition.  These two types
are not the same size on all architectures, and cause build failures on
systems where the types mismatch.

This commit changes the original function definition of lostCallback to
use the type C.uint64_t instead of C.ulong for the second argument, to
maintain compatibility across architectures.