elf: Apply same kern_version.SUBLEVEL clamping as done by kernel
When SUBLEVEL of the kernel version approaches 255, it is clamped to 255
to avoid impacting the PATCHLEVEL. This is introduced in different versions
depending on the branch. For example, for 4.19 kernel see the following
commit: https://git.kernel.org/pub/scm/linux/kernel/git/stable/linux.git/commit/?id=a256aac5b7000

In order to avoid failing the kernel version check in the BPF syscall, the
library must apply the same clamping.

Signed-off-by: David Verbeiren <david.verbeiren@tessares.net>