[bcc] Support for new `bcc_func_load` signature.
https://github.com/iovisor/bcc/commit/ffff0edc00ad249cffbf44d855b15020cc968536
a new parameter was added to control the attach type.
Subsequently, in
https://github.com/iovisor/bcc/commit/815d1b84828c02ce10e1ea5163aede6b5786ba38
, `-1` was used to signal the default behaviour.

This change adds this new parameter to the call to `C.bcc_func_load`
when LIBBCC_VERSION_CODE is greater or equal than 0.25.0, the next version to
include the aforementioned change.

Test plan:
Against v0.25.0 is being tested as part of this project CI.

also ran against the current v0.20.0 and ran CI: https://github.com/chantra/gobpf/runs/7436803048